requirements.txt
pandas
numpy
faker
openpyxl
click
rich
pyyaml